backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
pydantic = "^1.9.1"
python-dotenv = "^0.17.0"
poetry-dotenv-plugin = "^0.1.0-alpha.2"
peewee = "^3.15.1"
python-i18n = "^0.3.9"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
